Update for Xcode 7.3
Just remove __weak because in MRC.
Quote answer from Apple: http://stackoverflow.com/a/36148123/3796488

This issue behaves as intended based on the following: We are in the process of implementing weak references in all Objective-C language modes. Since “__weak” has historically been ignored in non-ARC (and non-GC) language modes, we’ve added this error to point out places where the semantics will change in the future. Please update your bug report to let us know if this is still an issue for you.